Add Vuln namespace for the original qualys importer
